backend/pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "precept-crm"
version = "0.1.0"
description = "Precept CRM — Django + Next.js CRM"
requires-python = ">=3.12"
dependencies = [
    "django>=5.1,<5.2",
    "django-ninja>=1.3,<2",
    "django-ninja-jwt>=5.3,<6",
    "psycopg[binary]>=3.2,<4",
    "django-redis>=5.4,<6",
    "celery[redis]>=5.4,<6",
    "django-celery-beat>=2.6,<3",
    "channels[daphne]>=4.1,<5",
    "channels-redis>=4.2,<5",
    "django-cors-headers>=4.4,<5",
    "Pillow>=10.4,<11",
    "beautifulsoup4>=4.12,<5",
    "phonenumbers>=8.13,<9",
    "twilio>=8.5,<9",
    "gunicorn>=22,<23",
]

[project.optional-dependencies]
dev = [
    "pytest>=8,<9",
    "pytest-django>=4.8,<5",
    "pytest-asyncio>=0.23,<1",
    "factory-boy>=3.3,<4",
    "ruff>=0.5,<1",
    "ipdb",
]

[tool.ruff]
line-length = 110
target-version = "py312"

[tool.ruff.lint]
select = ["F", "E", "W", "I", "UP", "B", "RUF"]

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "config.settings.test"
python_files = ["tests.py", "test_*.py"]
